FIPS-2159/FIPS-2744 :: dependency from formula to product component
Update ProductCmptGeneration and Expression to add dependency from
within a formula expression to a product component when asked for
dependent objects.

Added productComponent to QualifierNode to get the whole IProductCmpt
instead of only the runtime ID.

This state does not work with refactoring because the currently created
dependency detail implementation does not support the refactoring.
This will be implemented in the next task!

Change-Id: I3e069a9ee5690ea28e649b831926c13cf3b032dd