Bugfix: Do not set rage as valueSet for lower/upper/step control
The valueSet should only be provided to the control factory if the
expected value should be one of this value. In case of a range this will
not change anything according to the control. But our (not so smart)
detection of "default value control" would guess it is a default value
if there is a valueSet provided and would paint "<no default value>" in
range edit dialog. Resolution: do not provide a the valueSet because it
is wrong anyway.

Change-Id: Ib6f82bf9f24e592dc07c0111c8970fe3a24d4af4