Bugfix: Remove some unnecessary calls of bindingContext.updateUI
Calling bindingContext.updateUI leads to a bug in windows/eclipse 4
where the focus in a text control jumps to first position. The exact
point seems to be in the update of the error marker decoration image but
only when called via bindingContext.updateUI. As long as we do not
exactly know where the bug comes from, we just remove some unused
updateUI calls. Anyhow the updateUI method is only a good idea when
creating a new view (editor/dialog) and should not be necessary to
update any data while editing something.

Change-Id: Icc40996c1312655e7bea9e088b28523e070c8767