FIPS-4557/FIPS-4561 :: finish operation for infer template
Commit got a little big... next to obvious changes I did the following:
 * Subclass Operation of NewProductCmptWizard. First I also wanted to
separate the whole PMO but that would lead to a huge redesign.
 * Refactor ContentChangeEvent to have the ability to merge two
contentChangeEvents better while queueing them in a buld change
(rundAndQueueChangeEvents). That was needed to have the structural
change in product editor after infering the template
 * ValueHolders got ability to be copied and got concrete
IAttributeValue as Parent - IIpsObjectPart was far to generatl and not
useful
 * IValue is now comparable, comparator in ValueHolder now simpler

Change-Id: Ic22bf1d4f6c65f34e7d3db2ae5c845ddf233d9e5