FIPS-4483 :: fix for clean up runAndQueueChangeEvents
The new run method with busy indicator does not run in UI/display thread. That leads to exceptions when the executor expects the display thread.

Further more it is not usefull in combination with wizards. In wizards you should always call getContainer().run(...) and activating the progress bar by calling setNeeProgressBar(true)

Change-Id: I5acc36da059fa22a6aaede12dba79d1d65ecbd64